Navbar Rework + Pages Rework
