+++ resources/js/Pages/Payroll/Admin/Compensations.tsx
