+++ resources/js/Pages/Payroll/Auth/Register.tsx
