+++ resources/js/Pages/Payroll/Auth/VerifyEmail.tsx
