+++ resources/js/Pages/Payroll/Employee/Salary.tsx
