+++ resources/js/Pages/Profile/Edit.tsx
@@ -1,4 +1,4 @@
-import AuthenticatedLayout from "@/Layouts/AdminAuthenticatedLayout";
+import AuthenticatedLayout from "@/Layouts/AuthenticatedLayoutAdmin";
 import { PageProps } from "@/types";
 import { Head } from "@inertiajs/react";
 import DeleteUserForm from "./Partials/DeleteUserForm";
@@ -10,13 +10,7 @@ export default function Edit({
     status,
 }: PageProps<{ mustVerifyEmail: boolean; status?: string }>) {
     return (
-        <AuthenticatedLayout
-            header={
-                <h2 className="text-xl font-semibold leading-tight text-gray-800">
-                    Profile
-                </h2>
-            }
-        >
+        <AuthenticatedLayout>
             <Head title="Profile" />
 
             <div className="py-12">
